modularize test-matrix
